// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Role {
  id: number;
  company: string;
  logo?: (number | null) | Media;
  /**
   * External logo URL (used if no upload)
   */
  logoUrl?: string | null;
  /**
   * Overall summary for this company
   */
  summary?: string | null;
  /**
   * Roles held at this company, newest first
   */
  positions: {
    title: string;
    startDate: string;
    /**
     * Leave empty for current role
     */
    endDate?: string | null;
    id?: string | null;
  }[];
  keyWins?:
    | {
        win: string;
        id?: string | null;
      }[]
    | null;
  sortOrder?: number | null;
  updatedAt: string;
  createdAt: string;
}